+++ includes/integrations/views/integrations.php
@@ -41,7 +41,7 @@ function _nl4wp_integrations_table( $integrations ) {
 				</td>
 				<td class="desc">
 					<?php
-                    _e($integration->description, 'newsletter-for-wp');
+                   _e($integration->description, 'newsletter-for-wp');
                     ?>
 				</td>
 			</tr>